Move legacy expression code (WIP)
