+++ locale/en_US/LC_MESSAGES/django.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-08-22 19:27+0000\n"
+"POT-Creation-Date: 2019-08-28 22:40+0000\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -2460,9 +2460,6 @@ msgid "When this ruleset was originally created"
 msgstr ""
 
 msgid "When this ruleset was last modified"
-msgstr ""
-
-msgid "No subscribers to this event"
 msgstr ""
 
 msgid "When this action was originally created"
