+++ locale/fr/LC_MESSAGES/django.po
@@ -12,7 +12,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: TextIt\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-08-22 19:27+0000\n"
+"POT-Creation-Date: 2019-08-28 22:40+0000\n"
 "PO-Revision-Date: 2014-09-24 19:57+0000\n"
 "Last-Translator: TextIt <info@textit.in>\n"
 "Language-Team: French (http://www.transifex.com/projects/p/textit/language/fr/)\n"
@@ -3516,10 +3516,6 @@ msgstr "Quand ce jeu de règles a été créé à l'origine"
 
 msgid "When this ruleset was last modified"
 msgstr "Quand ce jeu de règles a été mise à jour"
-
-#, fuzzy
-msgid "No subscribers to this event"
-msgstr "Le numéro de téléphone pour cet appel"
 
 msgid "When this action was originally created"
 msgstr "Lorsque cette action a été créé à l'origine"
@@ -11262,6 +11258,10 @@ msgstr "Choisissez le moment où vous desirer que le flux démarre. "
 #, fuzzy
 msgid "Facebook Channel"
 msgstr "Supprimer téléphone"
+
+#, fuzzy
+#~ msgid "No subscribers to this event"
+#~ msgstr "Le numéro de téléphone pour cet appel"
 
 #, fuzzy
 #~| msgid "Channel Event"
