+++ locale/pt_BR/LC_MESSAGES/django.po
@@ -12,7 +12,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: TextIt\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-08-22 19:27+0000\n"
+"POT-Creation-Date: 2019-08-28 22:40+0000\n"
 "PO-Revision-Date: 2014-04-14 20:37+0000\n"
 "Last-Translator: produtos <produtos.ti@takenet.com.br>\n"
 "Language-Team: Portuguese (Brazil) (http://www.transifex.com/projects/p/textit/language/pt_BR/)\n"
@@ -3404,10 +3404,6 @@ msgstr "Quando este conjunto de regras foi originalmente criado"
 
 msgid "When this ruleset was last modified"
 msgstr "Quando este conjunto de regras foi modificado pela última vez"
-
-#, fuzzy
-msgid "No subscribers to this event"
-msgstr "O número de telefone para esta chamada"
 
 msgid "When this action was originally created"
 msgstr "Quando essa ação foi originalmente criada"
@@ -10815,6 +10811,10 @@ msgstr "Escolha quando você gostaria de iniciar o fluxo."
 #, fuzzy
 msgid "Facebook Channel"
 msgstr "Remover telefone"
+
+#, fuzzy
+#~ msgid "No subscribers to this event"
+#~ msgstr "O número de telefone para esta chamada"
 
 #, fuzzy
 #~| msgid "Channel Event"
