+++ temba/flows/legacy/migrations.py
@@ -8,7 +8,6 @@ from temba.contacts.models import ContactField, ContactGroup
 from temba.flows.models import Flow, RuleSet
 from temba.msgs.models import Label
 from temba.utils import json
-from temba.utils.expressions import migrate_template
 from temba.utils.languages import iso6392_to_iso6393
 
 from .definition import (
@@ -24,6 +23,7 @@ from .definition import (
     TriggerFlowAction,
     VariableContactAction,
 )
+from .expressions import migrate_v7_template
 
 
 def migrate_to_version_11_12(json_flow, flow=None):
@@ -1086,7 +1086,7 @@ def migrate_to_version_8(json_flow, flow=None):
 
     def migrate_node(node):
         if isinstance(node, str):
-            return migrate_template(node)
+            return migrate_v7_template(node)
         if isinstance(node, list):
             for n in range(len(node)):
                 node[n] = migrate_node(node[n])
